requirements.txt
numpy
pandas
openpyxl
tinkoff-investments
playwright
bs4
dotenv
aiohttp
async_lru